moveonly: add mp/type-decay.h
