+++ include/mp/type-decay.h
@@ -1,0 +1,38 @@
+// Copyright (c) 2025 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef MP_PROXY_TYPE_DECAY_H
+#define MP_PROXY_TYPE_DECAY_H
+
+#include <mp/util.h>
+
+namespace mp {
+template <typename LocalType, typename Value, typename Output>
+void CustomBuildField(TypeList<const LocalType>,
+    Priority<0>,
+    InvokeContext& invoke_context,
+    Value&& value,
+    Output&& output)
+{
+    BuildField(TypeList<LocalType>(), invoke_context, output, std::forward<Value>(value));
+}
+
+template <typename LocalType, typename Value, typename Output>
+void CustomBuildField(TypeList<LocalType&>, Priority<0>, InvokeContext& invoke_context, Value&& value, Output&& output)
+{
+    BuildField(TypeList<LocalType>(), invoke_context, output, std::forward<Value>(value));
+}
+
+template <typename LocalType, typename Value, typename Output>
+void CustomBuildField(TypeList<LocalType&&>,
+    Priority<0>,
+    InvokeContext& invoke_context,
+    Value&& value,
+    Output&& output)
+{
+    BuildField(TypeList<LocalType>(), invoke_context, output, std::forward<Value>(value));
+}
+} // namespace mp
+
+#endif // MP_PROXY_TYPE_DECAY_H
